+++ packages/chaingraph-types/src/node/id-generate.ts
@@ -6,10 +6,11 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-import { v7 as uuidv7 } from 'uuid'
+import { customAlphabet } from 'nanoid'
+import { nolookalikes } from 'nanoid-dictionary'
 
-export function generatePortID(): string {
-  // return `PO${customAlphabet(nolookalikes, 16)()}`
+export function generatePortID(propertyKey: string = ''): string {
+  return `${propertyKey}:PO${customAlphabet(nolookalikes, 16)()}`
   // generate sortable UUID for now
-  return uuidv7()
+  // return uuidv7()
 }